+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Remove `-`from breadcrumb.
+
 ## [1.3.1] - 2020-06-03
 
 ### Changed