+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `extraData` when `productOriginVtex` is `true`.
+
 ## [1.6.0] - 2020-06-18
 
 ### Added